+++ gcovr/tests/test_gcovr.py
@@ -57,6 +57,18 @@ skip_clean = None
 CC = os.path.split(env["CC"])[1]
 IS_CLANG = True if CC.startswith("clang") else False
 
+IS_WINDOWS = platform.system() == "Windows"
+if IS_WINDOWS:
+    import win32api
+    import string
+
+    used_drives = win32api.GetLogicalDriveStrings().split("\0")
+    sys.stdout.write(f"Used drives: {used_drives}")
+    free_drives = list(set(string.ascii_uppercase) - set(used_drives))
+    sys.stdout.write(f"Free drives: {free_drives}")
+    assert free_drives, "Must have at least one free drive letter"
+    env["GCOVR_TEST_DRIVE_WINDOWS"] = f"{free_drives[0]}:"
+
 CC_REFERENCE = env.get("CC_REFERENCE", CC)
 
 REFERENCE_DIRS = []
@@ -192,8 +204,6 @@ KNOWN_FORMATS = [
 def pytest_generate_tests(metafunc):
     """generate a list of all available integration tests."""
 
-    is_windows = platform.system() == "Windows"
-
     global skip_clean
     skip_clean = metafunc.config.getoption("skip_clean")
     generate_reference = metafunc.config.getoption("generate_reference")
@@ -249,14 +259,18 @@ def pytest_generate_tests(metafunc):
                 continue
 
             marks = [
+                pytest.mark.skipif(
+                    name == "simple1-drive-subst" and not IS_WINDOWS,
+                    reason="drive substitution only available on windows",
+                ),
                 pytest.mark.xfail(
                     name == "exclude-throw-branches"
                     and format == "html"
-                    and is_windows,
+                    and IS_WINDOWS,
                     reason="branch coverage details seem to be platform-dependent",
                 ),
                 pytest.mark.xfail(
-                    name == "rounding" and is_windows,
+                    name == "rounding" and IS_WINDOWS,
                     reason="branch coverage seem to be platform-dependent",
                 ),
                 pytest.mark.xfail(